renamed to FilterViewProvider.ts
